Use 64-bit registers for memory sizes

+++ src/wasmjit/runtime.h
@@ -95,8 +95,8 @@ struct Store {
 		size_t n_elts;
 		struct MemInst {
 			char *data;
-			uint32_t size;
-			uint32_t max; /* max of 0 means no max */
+			size_t size;
+			size_t max; /* max of 0 means no max */
 		} *elts;
 	} mems;
 	struct Addrs startfuncs;
